==> notebooks/03 - joining master and weedmaps.ipynb <==
import pandas as pd
from collections import Counter
import numpy as np
import glob
import re
import json

licenses = pd.read_csv("..//data//searchResults_active_retailers.csv")
weedmaps_not_joined = pd.read_csv("..//data//store_clean.csv")
check_cols = ["adult_use_cultivation", "adult_use_nonstorefront", 
              "adult_use_retail", "distributor", "medical_cultivation", 
              "medical_nonstorefront", "medical_retail", "microbusiness"]

len(weedmaps_not_joined)

Counter(licenses["License Type"])

weedmaps_not_joined.shape

## replace hyphens with nothing

licenses['License Number'] = licenses['License Number'].str.replace('-', '')

for col in check_cols:
    if weedmaps_not_joined[col].dtype != "float":
        weedmaps_not_joined[col] = weedmaps_not_joined[col].str.upper()
        weedmaps_not_joined[col] = weedmaps_not_joined[col].str.replace('-', '')

## Explore join statistics


for i in check_cols:
    try:
        merge = pd.merge(weedmaps_not_joined, 
                         licenses, 
                         left_on = i,
                         right_on = "License Number",
                         how = "outer",
                         indicator = True
        )
        print(Counter(merge["_merge"])["both"], i)
        stats = Counter(merge[merge["_merge"] == 'both']["License Type"])
        for x in stats:
            print(x, stats[x])
        print()
    except:
        print("failed", i)


| . | Adult use cultivation | Adult use nonstorefront | Adult use retail | Distributor | medical cultivation | medical nonstorefront | medical retail | microbusiness |
|---|---|---|---|---|---|---|---|---|
|retailer nonstorefront temp | 0 | 183 | 7 | 0 | 0 | 105 | 4 | 0 |
retailer temporary | 0 | 40 | 166 | 0 | 0 | 5 | 71 | 0
retailer | 0 | 0 | 6 | 0 | 0 | 0 | 6 | 0
microbusiness temporary | 0 | 0 | 4 | 0 | 0 | 0 | 1 | 25 |
distributor temporary | 0 | 40 | 1 | 1 | 0 | 5 | 1 | 0 |


## Actual join

import glob
sum([True for x in glob.glob("..\\data\\*") if "weedmaps.db" in x])

'''
Creates a new dataframe (weedmaps_joined) which contains the joins of weedmaps stores
that had a license number in the licenses dataframe. The license numbers are joined in 
the order that they appear on check_cols.

Create another dataframe (weedmaps_not_joined) which contains all the weedmap stores that
had licenses that could not be joined to the licenses dataframe. 
'''

#weedmaps_not_joined = weedmaps
licenses_joined = pd.DataFrame()
licenses_not_joined = licenses.copy()
c = 0
for i in check_cols:
    try:
        join_on_i = pd.merge(weedmaps_not_joined, licenses, left_on = i, right_on = 'License Number', how = 'inner')
        print ("joined:", join_on_i.shape[0], ' on:', i)
        c += join_on_i.shape[0]
        
        licenses_not_joined = licenses_not_joined[~licenses_not_joined["License Number"].isin(join_on_i["License Number"])]
        #weedmaps_not_joined = weedmaps_not_joined[~weedmaps_not_joined[i].isin(join_on_i['License Number'])]
        licenses_joined = pd.concat([licenses_joined, join_on_i])
        weedmaps_not_joined = weedmaps_not_joined[~weedmaps_not_joined[i].isin(join_on_i[i])]
        
        
    except:
        print("none for",i)



latent_data_structure = {}
for license in licenses_joined["License Number"].unique():
    latent_data_structure.update({license: licenses_joined[licenses_joined["License Number"] == license].id.tolist()})
    
for license in licenses_not_joined["License Number"].unique():
    if licenses_not_joined[licenses_not_joined["License Number"] == license]["License Type"].iloc[0] in ['Cannabis - Retailer Temporary License', 'Cannabis - Retailer Nonstorefront Temporary License']:
        latent_data_structure.update({license: []})

with open("..//data//latent.json", "w") as f:
    json.dump(latent_data_structure, f)

Previously, we had 220/?1168 stores join. Now we have 382 / 942 for v1

409 / 942

v2:

len(licenses_joined["License Number"].unique())

len(licenses_not_joined["License Number"].unique())

len(licenses["License Number"].unique())

weedmaps_not_joined.shape

licenses_not_joined.phone = licenses_not_joined.phone.astype(np.int64, errors = "ignore")
#licenses_not_joined = licenses_not_joined.reset_index(drop = True)
#licenses_not_joined["master_index"] = licenses_not_joined.index
licenses_not_joined.to_csv("..//data//licenses_not_joined.csv", index = False)

#licenses_joined.phone = licenses_joined.phone.astype(np.int64, errors = "ignore")
licenses_joined.to_csv('..//data//licenses_joined.csv', index = False)

pd.crosstab(licenses_joined["license_type"], licenses_joined["Adult-Use/Medicinal"])

c
