==> notebooks/04 - creating matches.ipynb <==
import pandas as pd
import numpy as np
import re
import nltk
import json

- is there many licenses that have conflicing 
- create a "licenses_not_joined" file
 - collect statistics on how many licenses joined correctly with the adult-use / medicinal column.

def fix_nums(s):
    if pd.isnull(s):
        return np.nan
    else:
        return s.replace(".0", "")

### pick up where we left off. give mateo a list of already joined things. rank all of the stores.

left_to_join = json.load(open("..//data//latent.json", "r"))
left_to_join = [x[0] for x in list(filter(lambda x: x[1] == 0, {x: len(left_to_join[x]) for x in left_to_join}.items()))]

master_list = pd.read_csv("..//data/searchResultsClean.csv", low_memory = False)
master_list.zip_code = master_list.zip_code.fillna(np.nan).astype(str).apply(lambda x: fix_nums(x))
master_list.phone = master_list.phone.fillna(np.nan).astype(str).apply(lambda x: fix_nums(x))

weedmaps = pd.read_csv("..//data//store.csv")

Remove hyphen from phone number in weedmaps:

def remove_symbol(row, col, left, right):
    if pd.isnull(row[col]):
        return ""
    return re.sub(left, right, row[col], flags = re.I)

weedmaps.phone = weedmaps.apply(lambda row: remove_symbol(row, "phone", "\(|\)|-|\s|\.|\+", ""), axis = 1)

def search_column(row, col, string):
    if pd.isnull(row[col]):
        return False
    elif re.search(string, row[col], flags = re.I) != None:
        return True
    else:
        return False

def update_matches(test, matches, num):
    
    for match in test.id:
        if match not in matches:
            matches.update({match: num})
        else:
            matches[match] += num
    
    return matches

def suggest_stores(master_list, weedmaps):
    
    tmp = 0
    ranked_indexes = {}
    unique_cities = weedmaps.city.unique()
    unique_phones = weedmaps.phone.unique()
    unique_website = [x.strip() for x in weedmaps.website.unique() if not pd.isnull(x)]
    unique_emails = [x.strip() for x in weedmaps.email.unique() if not pd.isnull(x)]
    
    for unjoined_store in left_to_join:
        
        comparison = master_list[master_list["License_no_dash"] == unjoined_store]
        matches = {}
        
        #print(comparison)
        #print(comparison.website.iloc[0])
        #print(comparison.email.iloc[0])
        #print(comparison.phone.iloc[0])
        #print(comparison.zip_code.iloc[0])
        #print(comparison.city.iloc[0])
        #print(comparison.company_name.iloc[0])

        #print()
        # compare website
        if not pd.isnull(comparison.website.iloc[0]):
            best_site = sorted({x: nltk.edit_distance(comparison.website.iloc[0], str(x)) for x in unique_website}.items(), key = lambda x: x[1])[0]
            #print(best_site)
            if best_site[1] <= 2:
                test = weedmaps[weedmaps.apply(lambda row: search_column(row, "website", best_site[0]), axis = 1)]
                matches = update_matches(test, matches, 3)
        
        
        # compare email
        if not pd.isnull(comparison.email.iloc[0]):
            best_email = sorted({x: nltk.edit_distance(comparison.email.iloc[0], str(x)) for x in unique_emails}.items(), key = lambda x: x[1])[0]
            #print(best_email)
            if best_email[1] <= 2:
                test = weedmaps[weedmaps.apply(lambda row: search_column(row, "email", best_email[0]), axis = 1)]
                matches = update_matches(test, matches, 3)
        
        
        # compare phone
        if not pd.isnull(comparison.phone.iloc[0]):
            best_phone = sorted({x: nltk.edit_distance(comparison.phone.iloc[0], x) for x in unique_phones}.items(), key = lambda x: x[1])[0]
            #print(best_phone)
            if best_phone[1] <= 1:
                test = weedmaps[weedmaps.apply(lambda row: search_column(row, "phone", best_phone[0]), axis = 1)]
                matches = update_matches(test, matches, 3)
        
        
        # compare city
        if not pd.isnull(comparison.city.iloc[0]):
            best_city = sorted({x: nltk.edit_distance(comparison.city.iloc[0], x) for x in unique_cities}.items(), key = lambda x: x[1])[0][0]
            #print(best_city)
            test = weedmaps[weedmaps.apply(lambda row: search_column(row, "city", best_city), axis = 1)]    
            matches = update_matches(test, matches, 1)
            
            
        # compare names
        if not pd.isnull(comparison.company_name.iloc[0]):
            best_name = sorted({x: nltk.edit_distance(comparison.company_name.iloc[0], x) for x in unique_cities}.items(), key = lambda x: x[1])[0][0]
            #print(best_name)
            test = weedmaps[weedmaps.apply(lambda row: search_column(row, "name", best_city), axis = 1)]
            matches = update_matches(test, matches, 1)  
        
        
        # compare zip
        test = weedmaps[weedmaps.apply(lambda row: search_column(row, "zip_code", comparison.zip_code.iloc[0]), axis = 1)]
        #print(test)
        matches = update_matches(test, matches, 1)
        
        ranked_indexes.update({unjoined_store: sorted(matches.items(), key = lambda x: x[1], reverse = True)})
        
    return ranked_indexes

all_matches = suggest_stores(master_list, weedmaps)

with open("..//data//matches.json", "w") as f:
    json.dump(all_matches, f)

.

.

.

.

.

.

.

.

.

.

.

.

.

## match email

left_to_merge.head()

master_list[master_list.apply(lambda row: search_column(row, "Business Contact Information", "vhhc420@gmail.com"), axis = 1)]

## match phone

master_list[master_list.apply(lambda row: search_column(row, "Business Contact Information", "7076525018"), axis = 1)]

## match zip

master_list[master_list.apply(lambda row: search_column(row, "zip_code", "94903"), axis = 1)]

## match city

master_list[master_list.apply(lambda row: search_column(row, "Premise Address", "Marin"), axis = 1)]

def suggest_stores(left_to_merge, master_list):
    
    tmp = 0
    ranked_indexes = {}
    unique_cities = master_list.city.unique()

    for index in left_to_merge.id:
        comparison = left_to_merge[left_to_merge.id == index]
        matches = {}

        # check phone
        #print(ranked_indexes)
        print(comparison.phone.iloc[0])
        print(comparison.zip_code.iloc[0])
        print(comparison.city.iloc[0])
        #print()
        
        test = master_list[master_list.apply(lambda row: search_column(row, "Business Contact Information", comparison.phone.iloc[0]), axis = 1)]    
        matches = update_matches(test, matches, 2)

        # check zip
        test = master_list[master_list.apply(lambda row: search_column(row, "zip_code", comparison.zip_code.iloc[0]), axis = 1)]    
        matches = update_matches(test, matches, 1)

        # check city
        # first find closest city wrt levenschtein distance
        best_city = sorted({x: nltk.edit_distance(comparison.city.iloc[0].lower(), x) for x in unique_cities}.items(), key = lambda x: x[1])[0][0]
        #print(best_city)
        # then find those indices
        test = master_list[master_list.apply(lambda row: search_column(row, "city", best_city), axis = 1)]    
        matches = update_matches(test, matches, 1)

        # sort matches
        ranked_indexes.update({index: sorted(matches.items(), key = lambda x: x[1], reverse = True)})
        
        tmp += 1
        if tmp > 20:
            break

    return ranked_indexes
